Refactor Countdown Activity layout and enhance bottom navigation
This commit refactors the `activity_countdown.xml` layout to improve the structure and appearance. It updates the buttons text in `activity_main.xml` to `Countdown` and `Calculate` and it introduces a `ConstraintLayout` in `activity_countdown.xml` to replace `LinearLayout` and adds a `ScrollView` with fillViewport. It updates the layout with `header_countdown.xml` for header section and adds `bottomNav` for bottom navigation. Additionally, the bottom navigation bar is enhanced by setting background color to `colorSurface`, adding elevation, updating colors of the icon and text and using ripple color.