Refactor UI and improve icons, update deadline list display
This commit refactors the user interface, improves the appearance and functionality of icons, and enhances the display of the deadline list. It introduces new icons for various actions and states, and updates the `MainActivity` layout to use new icons. Additionally, it refactors the deadline list to display deadlines in a more organized and visually appealing manner, updates `DeadlineAdapter` and `DeadlineAdapterModern` to use new icons and `Deadline` class. The colors of the text in the deadline list are also changed to make it more consistent.
Specifically, the following changes were made:
- Added new icons for backspace, calendar clock, clear all, close, cycle, event, hourglass, and radio button checked.
- Updated the "Select Date" and "Countdown" buttons to use new icons.
- Updated the "Saved Deadlines" label color to `onBackgroundDisabled`.
- Updated `deadlineRecyclerView` properties.
- Updated `clearAll` button to use `clear_all` icon.
- Updated `DeadlineAdapter` to use `radio_button_checked` icon.
- Replaced the old deadline list with a `RecyclerView` and `MaterialCardView`.
- Added `DeadlineAdapterModern` to use new deadline view.
- Updated `item_deadline` and `item_deadline_modern` layouts to use new icons.
- Changed refresh icon in `widget_countdown` to `cycle` icon.
- Adjusted the text size and styling of deadline items in the `item_deadline_modern` layout.
- Updated the color definitions.