Refactor widget layout and update text sizes
This commit refactors the widget layout and updates text sizes for improved readability. It adjusts the padding in `widget_countdown.xml` to `8dp` and the top and bottom padding to `2dp`. The progress bar width is set to match parent and height to match parent. Additionally, the title size in `WidgetLayoutManager.kt` is changed from `16f` to `22f` , and from `22f` to `24f`, timer size from `26f` to `34f` and from `30f` to `40f`.